requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto==2.49.0
boto3==1.21.2
botocore==1.24.2
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-annoying==0.10.6
django-appconf==1.0.5
django-authtools==1.7.0
django-background-tasks==1.2.5
django-bootstrap3==21.2
django-braces==1.15.0
django-compat==1.0.15
django-compressor==3.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-debug-toolbar-django13==0.8.4
django-debugtools==2.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-imperavi==0.2.3
django-infinite-scroll-pagination==1.2.1
django-jet==1.0.8
django-jsonstore==0.5.0
django-material==1.10.0
django-rest-framework==0.1.0
django-storages==1.12.3
django-tinymce==3.4.0
django-viewflow==1.10.1
djangorestframework==3.13.1
dnspython==2.2.0
docutils==0.18.1
elastic-transport==8.0.1
elasticsearch==8.0.0
elasticsearch-dsl==7.4.0
email-validator==1.1.3
et-xmlfile==1.1.0
filelock==3.6.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httplib2==0.20.4
httpx==0.18.2
idna==3.3
jmespath==0.9.3
ldif3==3.2.2
newrelic==7.4.0.172
numpy==1.22.2
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.0
pluggy==1.0.0
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-card-me==0.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyuploadcare==3.0.0
rcssmin==1.1.0
requests==2.27.1
rfc3986==1.5.0
rjsmin==1.2.0
rsa==4.8
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
surlex==0.2.0
toml==0.10.2
tox==3.24.5
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
